Remove public compression package, and move to gRPC.
