+++ exporter/opencensusexporter/factory_test.go
@@ -24,7 +24,6 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector/component/componenttest"
-	"github.com/open-telemetry/opentelemetry-collector/compression"
 	"github.com/open-telemetry/opentelemetry-collector/config/configcheck"
 	"github.com/open-telemetry/opentelemetry-collector/config/configgrpc"
 	"github.com/open-telemetry/opentelemetry-collector/exporter/exportertest"
@@ -121,7 +120,7 @@ func TestCreateTraceExporter(t *testing.T) {
 			config: Config{
 				GRPCSettings: configgrpc.GRPCSettings{
 					Endpoint:    rcvCfg.Endpoint,
-					Compression: compression.Gzip,
+					Compression: configgrpc.CompressionGzip,
 				},
 			},
 		},
